[PVR] Fix some potential data narrowing issues by using proper data types. Replace still needed data narrowing 'static_cast' occurences with 'KODI::UTILS::Narrow' to make them searchable and to crash on data mismatch in debug builds.

+++ xbmc/pvr/channels/PVRChannelGroupsContainer.h
@@ -168,7 +168,7 @@ public:
    * @param providerId The providerId.
    * @return The total count of matching channels.
    */
-  unsigned int GetChannelCountByProvider(bool isRadio, int clientId, int providerId) const;
+  size_t GetChannelCountByProvider(bool isRadio, int clientId, int providerId) const;
 
   /*!
    * @brief Erase stale texture db entries and image files.
